requirements.txt
# See to Reach, Feel to Insert - 统一环境依赖
# 经过验证的配置，支持 DP (PyTorch) + SERL (JAX)
#
# 关键版本要求:
#   - Python 3.11
#   - NumPy 2.x (JAX 0.7.x 要求)
#   - diffusers 0.30+ (兼容 JAX 0.7.x)
#   - CUDA 12.x
#
# 安装: pip install -r requirements.txt

# ============================================
# PyTorch (CUDA 12.1)
# 注意: 需要从 PyTorch 官方 wheel 安装
# pip install torch torchvision --index-url https://download.pytorch.org/whl/cu121
# ============================================
# torch>=2.2.0
# torchvision>=0.17.0

# ============================================
# JAX (CUDA 12, 需要 NumPy 2.x)
# ============================================
jax[cuda12]>=0.7.0
flax>=0.8.0
orbax-checkpoint>=0.5.0
optax>=0.1.5
chex>=0.1.85
distrax>=0.1.2

# ============================================
# Diffusers (需要 0.30+ 兼容 JAX 0.7.x)
# ============================================
diffusers>=0.30.0
huggingface_hub>=0.25.0
accelerate>=0.20.0

# ============================================
# Hydra (DP 配置系统)
# ============================================
hydra-core>=1.2.0
omegaconf>=2.2.0

# ============================================
# TensorFlow (SERL 依赖)
# ============================================
tensorflow>=2.15.0
tensorflow_probability>=0.23.0
tf-keras

# ============================================
# 科学计算
# ============================================
numpy>=2.0
scipy>=1.11
einops>=0.6

# ============================================
# 图像处理
# ============================================
opencv-python>=4.6.0
imageio>=2.22
imageio-ffmpeg>=0.4
scikit-image>=0.19

# ============================================
# Gym/Gymnasium
# ============================================
gym>=0.26.0
gymnasium>=0.29.0

# ============================================
# 机器人接口
# ============================================
pyrealsense2>=2.51.0
pynput>=1.7.6
pyspacemouse>=1.0.0
requests>=2.28.0

# ============================================
# HIL-SERL
# ============================================
agentlace>=0.1.3
ml_collections>=0.1.0
moviepy>=1.0

# ============================================
# 日志和可视化
# ============================================
tensorboard>=2.10
wandb>=0.13
matplotlib>=3.6

# ============================================
# 数据存储
# ============================================
h5py>=3.7
zarr>=2.12
numcodecs>=0.10

# ============================================
# 工具
# ============================================
tqdm>=4.64
dill>=0.3.5
click>=8.0
pyyaml>=6.0
termcolor>=2.0
psutil>=5.9
natsort>=8.0
absl-py>=2.0
cloudpickle>=3.0
